requirements.txt
anyio==3.7.0
bcrypt==4.0.1
beanie==1.19.2
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.97.0
h11==0.14.0
idna==3.4
lazy-model==0.0.5
motor==3.1.2
passlib==1.7.4
pyasn1==0.5.0
pydantic==1.10.9
pymongo==4.3.3
python-decouple==3.8
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.6.3
uvicorn==0.22.0
